more test

+++ Capstone/Menu.cs
@@ -75,10 +75,10 @@ namespace Capstone
             {
                 Console.WriteLine("Please insert money in whole dollars($1, $2, $5, or $10)");
                 string currentMoneyProvidedString = Console.ReadLine();
-                decimal newCurrentMoneyProvided = int.Parse(currentMoneyProvidedString);
+                decimal newCurrentMoneyProvided = decimal.Parse(currentMoneyProvidedString);
                 currentBalance = currentBalance + newCurrentMoneyProvided;
                 Console.WriteLine("Current money provided: " + "$" + currentBalance);
-
+                
                 try
                 {
                     using (StreamWriter sw = new StreamWriter(outputFullPath, true))
